feat: provide database name in secret
The database name must be somehow unique across all namespace. So it can't be build solely on the `metadata.name`. But apart from hidden conventions, that's the only thing, the one creating the deployment knows. So it is better to provide the name.

Although the database name is not secret, we use the same resource (instead of a ConfigMap) to minimize API operations that may fail.