feat: provide ready to use remote base for releases
This commit changes the behavior of the release process.

Before each release, the image that will be created from the release is defined in a Kustomize release base. The base can be referenced as a remote base referring to the tag.

4 variants are provided:

- a Kustomize remote base without namespace definition to include in a `kustomization.yaml` that declares its own namespace: `https://github.com/SDA-SE/mongodb-operator//kustomize/bases/release-no-ns/?ref=<latest-tag>`
- a Kustomize remote base with namespace definition to include in a `kustomization.yaml` when the namespace `mongodb-operator` suits your needs: `https://github.com/SDA-SE/mongodb-operator//kustomize/bases/release/?ref=<latest-tag>`
- a rendered bundle without namespace: `kustomize/bundle-no-ns.yaml`
- a rendered bundle with namespace: `kustomize/bundle.yaml`

All variants require to define a secret named `mongodb-operator` with the key `mongodbConnectionString` providing the [connection string](https://www.mongodb.com/docs/manual/reference/connection-string/) for the MongoDB Operator to connect to the database.