feat(deployment): allow `patch` in addition to `update` for the MongoDB Operator
BREAKING CHANGE:

Due to a major update of the Operator SDK, the MongoDB Operator requires also `patch` privileges on resources that so far needed `update`. This change is covered for users who deploy with [the provided Kustomization](https://sda-se.github.io/mongodb-operator/deployment/#deployment-with-kustomize) but must be applied individually in case of a [manual deployment](https://sda-se.github.io/mongodb-operator/deployment/#manual-deployment).